+++ lib/scryglass.rb
@@ -1,10 +1,12 @@
 # frozen_string_literal: true
 require 'stringio'
+
 ## Bookkeeping and external tools:
 require "scryglass/version"
 require 'active_support/core_ext/object/blank' # This gives us `.present?` and `.blank?` # https://stackoverflow.com/questions/4648684/how-to-use-present-in-ruby-projects
 require 'io/console'
 require 'pp'
+require 'timeout'
 
 ## Refinements and sub-tools:
 require 'refinements/ansiless_string_refinement'