+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Added ability to distinguish genuine escape key presses
+
+## Changed
+
+- Now inputs check and screen redraws every 0.05 seconds even without user keypresses.
+
+## Removed
+
+- Temporarily removed record/playback functionality
+
 ## [1.0.1] - 2020-09-18
 
 ### Added